Time: 883 ms (29.91%), Space: 58 MB (44.78%) - LeetHub

+++ 1711-count-good-meals/1711-count-good-meals.cpp
@@ -11,7 +11,7 @@ public:
         for(int i=0;i<n;i++){
             
         for(int j=0;j<=21;j++){
-            int t=pow(2,j);
+            long long int t=pow(2,j);
             
             if(m.find(t-d[i])!=m.end())ans=(ans+m[t-d[i]])%mod;
